refactor: the shared build code becomes a package instead of a path out of one
Eight clients reached into `clients/tv-build` and `clients/tv-frame.vite.ts`
with a relative path. Neither is a package, so nothing recorded that
dependency and their imports resolved from wherever the path happened to
land — which is why moving one file broke `esbuild` resolution while the
typecheck still passed.

They are `@kroma/bundler` now, depended on by name: shell, rnw, tv-frame,
legacy-css, legacy-finalize, props-docs, polyfills-legacy, workbench-stub.
Declaring the package's own dependencies is what surfaced the debt — it
needed @kroma/ui, @kroma/build-info, postcss, lightningcss and the rest,
none of which any consumer had listed.

Two more traversals fixed on the way: shell.ts reached `@kroma/build-info`
through `../build-info/index.js` when the package has a name, and the root
vitest config imported props-docs and rnw by path.

Intra-package links go through the package's own name rather than relative
paths: Vite loads a config through Node ESM, which needs a `.ts` extension
on a relative import, and TypeScript rejects that extension without
`allowImportingTsExtensions`.

Verified: repo typecheck, 3066 tests, and the webos, tizen, desktop, kit
and web builds.

Still relative: the CLI scripts left in clients/tv-build (check-legacy,
stamp-version, perf-*, store-*), which are invoked by path from npm
scripts rather than imported.
