+++ clients/expo-build/metro-workspace.ts
@@ -69,7 +69,7 @@ function expoWorkspaceConfig(projectRoot, aliases = {}, ui = {}) {
     // `imports` and mirrored here because Metro does not read that field. Every
     // native client gets it for free: a shell should not have to know that the
     // kit refers to itself as `#ui`. The Vite half is `webResolve` in
-    // clients/tv-build/rnw.ts, and tsconfig.base.json carries the types.
+    // packages/bundler/src/rnw.ts, and tsconfig.base.json carries the types.
     '#ui': path.join(workspaceRoot, 'packages', 'ui', 'src'),
     ...aliases,
     'react-native': reactNative,
