+++ clients/kit/package.json
@@ -41,6 +41,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.28.0",
+    "@kroma/bundler": "workspace:*",
     "@react-native-tvos/config-tv": "^0.1.6",
     "@types/node": "^22.10.2",
     "@types/react": "~19.2.2",
