+++ clients/tizen/package.json
@@ -25,6 +25,7 @@
     "react-native-web": "^0.21.2"
   },
   "devDependencies": {
+    "@kroma/bundler": "workspace:*",
     "@tailwindcss/vite": "^4",
     "@types/node": "^22.10.2",
     "@types/react": "^19",
