+++ package.json
@@ -65,6 +65,7 @@
   "devDependencies": {
     "@biomejs/biome": "^2.5.5",
     "@csstools/postcss-cascade-layers": "^6.0.0",
+    "@kroma/bundler": "workspace:*",
     "@tailwindcss/vite": "^4.3.3",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.2",
