+++ packages/bundler/src/props-docs.ts
