+++ packages/bundler/src/workbench-stub.tsx
